+++ app.go
@@ -13,7 +13,7 @@ const (
 	// Default user agent for requests
 	userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36"
 
-	requestTimeout = 5 * time.Second
+	requestTimeout = 10 * time.Second
 
 	// RabbitMQ queue names:
 	queueProxySources         = "proxy_sources"