+++ database.go
@@ -8,8 +8,6 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 )
 
-const tableProxies = "proxies"
-
 type Proxy struct {
 	ID        int
 	Active    bool
@@ -22,33 +20,90 @@ type Proxy struct {
 	LastCheck time.Time
 }
 
-func proxyExistInDB(pool *pgxpool.Pool, u string) bool {
-	s, _, err := sq.Select("COUNT(*)").
-		From(tableProxies).
-		Where(sq.Eq{"url": u}).
+// SQLBuilder is a struct for sql statement.
+type SQLBuilder struct {
+	table string
+}
+
+// newSQLBuilder gets the table name and returns SQLBuilder.
+func newSQLBuilder(table string) SQLBuilder {
+	return SQLBuilder{table: table}
+}
+
+type WhereCondition map[string]interface{}
+
+// count gets WhereCondition and returns SELECT COUNT sql statement, sql args and an error.
+// count gets only "equal" (whereEq) and "greater or equal" (whereGtOrEq) conditions.
+func (s *SQLBuilder) count(whereEq, whereGtOrEq WhereCondition) (string, []interface{}, error) {
+	selectBuilder := sq.Select("COUNT(*)").From(s.table)
+
+	if len(whereEq) > 0 && len(whereGtOrEq) > 0 {
+		selectBuilder = selectBuilder.Where(sq.And{
+			sq.Eq(whereEq),
+			sq.GtOrEq(whereGtOrEq),
+		})
+	} else if len(whereEq) > 0 {
+		selectBuilder = selectBuilder.Where(sq.Eq(whereEq))
+	} else if len(whereGtOrEq) > 0 {
+		selectBuilder = selectBuilder.Where(sq.GtOrEq(whereGtOrEq))
+	}
+
+	return selectBuilder.PlaceholderFormat(sq.Dollar).ToSql()
+}
+
+// insert gets data and returns INSERT sql statement, sql args and an error.
+// Example data:
+//  map[string]interface{}{"column_name": "value"}
+func (s *SQLBuilder) insert(data map[string]interface{}) (string, []interface{}, error) {
+	if len(data) == 0 {
+		return "", nil, errDBInsertEmptyData
+	}
+
+	cols := make([]string, 0)
+	values := make([]interface{}, 0)
+	for k, v := range data {
+		cols = append(cols, k)
+		values = append(values, v)
+	}
+
+	return sq.
+		Insert(s.table).
+		Columns(cols...).
+		Values(values...).
+		PlaceholderFormat(sq.Dollar).
 		ToSql()
+}
+
+func proxyExistInDB(pool *pgxpool.Pool, hostPort string) (bool, error) {
+	sb := newSQLBuilder(tableProxies)
+	s, args, err := sb.count(WhereCondition{"url": hostPort}, WhereCondition{})
 	if err != nil {
-		return false
+		return false, err
 	}
 
 	var counter int
-	if err := pool.QueryRow(context.Background(), s).Scan(&counter); err != nil {
-		return false
+	if err = pool.QueryRow(context.Background(), s, args...).Scan(&counter); err != nil {
+		return false, err
 	}
 
-	if counter > 0 {
-		return true
+	if counter != 0 {
+		return true, nil
 	}
-	return false
+	return false, nil
 }
 
 func insertProxyIntoDB(pool *pgxpool.Pool, p Proxy) error {
-	s, args, err := sq.
-		Insert(tableProxies).
-		Columns("url", "active", "http", "https", "socks5", "anonymous", "created", "last_check").
-		Values(p.URL, p.Active, p.HTTP, p.HTTPS, p.SOCKS5, p.Anonymous, p.Created, p.LastCheck).
-		PlaceholderFormat(sq.Dollar).
-		ToSql()
+	sb := newSQLBuilder(tableProxies)
+	s, args, err := sb.insert(map[string]interface{}{
+		"url":        p.URL,
+		"active":     p.Active,
+		"http":       p.HTTP,
+		"https":      p.HTTPS,
+		"socks5":     p.SOCKS5,
+		"anonymous":  p.Anonymous,
+		"created":    p.Created,
+		"last_check": p.LastCheck,
+	})
 	if err != nil {
 		return err
 	}
@@ -85,20 +140,25 @@ func updateProxyInDB(pool *pgxpool.Pool, p Proxy) error {
 	return nil
 }
 
-func saveProxyToDB(pool *pgxpool.Pool, p Proxy) error {
-	exist := proxyExistInDB(pool, p.URL)
+func saveProxyToDB(pool *pgxpool.Pool, p Proxy) (successType string, err error) {
+	exist, err := proxyExistInDB(pool, p.URL)
+	if err != nil {
+		return "", err
+	}
 
 	if exist {
 		if err := updateProxyInDB(pool, p); err != nil {
-			return err
+			return "", err
 		}
+		successType = "update"
 	} else {
 		if err := insertProxyIntoDB(pool, p); err != nil {
-			return err
+			return "", err
 		}
+		successType = "insert"
 	}
 
-	return nil
+	return successType, nil
 }
 
 func changeProxyToInactive(pool *pgxpool.Pool, pURL string) error {