+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - added tests for findProxiesHostPort
 - added values to queriesForSearchEngine
 - added findProxySourcesFromDDG description
+- increased requests timeout
 
 ## 0.3.2 - 2021.08.14